pyproject.toml
[tool.poetry]
name = "GreyWolf_Difference_Generator"
version = "0.2.0"
description = "Gendiff"
authors = ["GreyGreyWolf"]
packages = [
  { include = "gen_diff" },
  ]


[tool.poetry.dependencies]
python = "^3.6"
pyyaml = "^5.3"


[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
pytest = "^5.3.5"
pytest-cov = "^2.8.1"
coverage = "^5.0.3"

[tool.poetry.scripts]
  gendiff = "gen_diff.scripts.gendiff:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
